✅ Add tests for snippets in the website
