+++ website/.gitignore
@@ -7,6 +7,7 @@
 # Generated files
 .docusaurus
 .cache-loader
+generated-tests
 
 # Misc
 .DS_Store
