+++ website/docs/tutorials/detect-race-conditions/Playgrounds.jsx
@@ -9,397 +9,23 @@ import {
 } from '@codesandbox/sandpack-react';
 import { atomDark } from '@codesandbox/sandpack-themes';
 import styles from './Playgrounds.module.css';
-
-// Does not queue at all
-const queueCodeV0 = `export function queue(fun) {
-  return fun;
-}`;
-
-// Only queue after last request
-const queueCodeV1 = `export function queue(fun) {
-  let lastQuery = Promise.resolve();
-  return (...args) => {
-    const currentQuery = fun(...args);
-    const returnedQuery = lastQuery.then(() => currentQuery);
-    lastQuery = currentQuery;
-    return returnedQuery;
-  };
-}`;
-
-// May still not queue enough due to early cleaning
-const queueCodeV2 = `export function queue(fun) {
-  let pastQueries = [];
-  return (...args) => {
-    const currentQuery = fun(...args);
-    const returnedQuery = Promise.all(pastQueries)
-      .finally(() => (pastQueries = []))
-      .then(() => currentQuery);
-    pastQueries.push(currentQuery);
-    return returnedQuery;
-  };
-}`;
-
-const queueCodeV3 = `export function queue(fun) {
-  let pastQueries = [];
-  return (...args) => {
-    const currentQuery = fun(...args);
-    const knownPastQueries = pastQueries;
-    const returnedQuery = Promise.all(pastQueries)
-      .finally(() => {
-        if (knownPastQueries === pastQueries)
-          pastQueries = [];
-      })
-      .then(() => currentQuery);
-    pastQueries = [...pastQueries, currentQuery];
-    return returnedQuery;
-  };
-}`;
-
-const queueCodeV4 = `export function queue(fun) {
-  let queryId = 0;
-  let lastQuery = null;
-  return (...args) => {
-    const selfQueryId = ++queryId;
-    if (lastQuery === null) {
-      lastQuery = fun(...args);
-    } else {
-      lastQuery = lastQuery
-        .then(
-          () => fun(...args),
-          () => fun(...args),
-        );
-    }
-    lastQuery
-      .finally(() => {
-        if (queryId === selfQueryId)
-          lastQuery = null;
-      });
-    return lastQuery;
-  };
-}`;
-
-const queueCodeV5 = `export function queue(fun) {
-  let pending = false;
-  let onDone = [];
-  function runNext() {
-    if (onDone.length === 0) {
-      pending = false;
-      return;
-    }
-    onDone.shift()();
-  }
-  return (...args) => {
-    if (!pending) {
-      pending = true;
-      const p = fun(...args);
-      p.then(runNext, runNext);
-      return p;
-    }
-    return new Promise((resolve, reject) => {
-      onDone.push(() => {
-        const p = fun(...args);
-        p.then(runNext, runNext);
-        p.then(resolve, reject);
-      });
-    });
-  };
-}`;
-
-const queueUnitSpecCode = `import {queue} from './queue.js';
-
-test('should resolve in call order', async () => {
-  // Arrange
-  const seenAnswers = [];
-  const call = jest.fn()
-    .mockImplementation(v => Promise.resolve(v));
-
-  // Act
-  const queued = queue(call);
-  await Promise.all([
-    queued(1).then(v => (seenAnswers.push(v))),
-    queued(2).then(v => (seenAnswers.push(v))),
-  ]);
-
-  // Assert
-  expect(seenAnswers).toEqual([1, 2]);
-})`;
-
-const queueBasicPBTSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), async (s) => {
-    // Arrange
-    const pendingQueries = [];
-    const seenAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-  
-    // Act
-    const queued = queue(s.scheduleFunction(call));
-    pendingQueries.push(queued(1).then(v => (seenAnswers.push(v))));
-    pendingQueries.push(queued(2).then(v => (seenAnswers.push(v))));
-    await s.waitFor(Promise.all(pendingQueries));
-  
-    // Assert
-    expect(seenAnswers).toEqual([1, 2]);
-  }))
-})`;
-
-const queueBasicPBTWaitAllSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), async (s) => {
-    // Arrange
-    const seenAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-  
-    // Act
-    const queued = queue(s.scheduleFunction(call));
-    queued(1).then(v => (seenAnswers.push(v)));
-    queued(2).then(v => (seenAnswers.push(v)));
-    await s.waitAll();
-  
-    // Assert
-    expect(seenAnswers).toEqual([1, 2]);
-  }))
-})`;
-
-const queueMoreThan2CallsPBTSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), fc.integer({min: 1, max: 10}), async (s, numCalls) => {
-    // Arrange
-    const pendingQueries = [];
-    const seenAnswers = [];
-    const expectedAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-  
-    // Act
-    const queued = queue(s.scheduleFunction(call));
-    for (let id = 0 ; id !== numCalls ; ++id) {
-      expectedAnswers.push(id);
-      pendingQueries.push(queued(id).then(v => (seenAnswers.push(v))));
-    }
-    await s.waitFor(Promise.all(pendingQueries));
-  
-    // Assert
-    expect(seenAnswers).toEqual(expectedAnswers);
-  }))
-})`;
-
-const queueBatchesAlternativePBTSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), fc.integer({min: 1, max: 10}), async (s, numCalls) => {
-    // Arrange
-    const pendingQueries = [];
-    const seenAnswers = [];
-    const expectedAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-  
-    // Act
-    const queued = queue(s.scheduleFunction(call));
-    for (let id = 0 ; id !== numCalls ; ++id) {
-      pendingQueries.push(
-        s.schedule(Promise.resolve(\`Fire the call for \${id}\`))
-          .then(() => {
-            expectedAnswers.push(id);
-            return queued(id);
-          })
-          .then(v => (seenAnswers.push(v)))
-      );
-    }
-    await s.waitFor(Promise.all(pendingQueries));
-  
-    // Assert
-    expect(seenAnswers).toEqual(expectedAnswers);
-  }))
-})`;
-
-const queueFromBatchesPBTSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), fc.array(fc.integer({min: 1, max: 10}), {minLength: 1}), async (s, batches) => {
-    // Arrange
-    const pendingQueries = [];
-    const seenAnswers = [];
-    const expectedAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-  
-    // Act
-    const queued = queue(s.scheduleFunction(call));
-    let lastId = 0;
-    const { task } = s.scheduleSequence(batches.map((batch, index) => {
-      return {
-        label: \`Fire batch #\${index + 1} (\${batch} calls)\`,
-        builder: async () => {
-          for (let id = 0 ; id !== batch ; ++id, ++lastId) {
-            expectedAnswers.push(lastId);
-            pendingQueries.push(queued(lastId).then(v => (seenAnswers.push(v))));
-          }
-        },
-      }
-    }));
-    await s.waitFor(task);
-    await s.waitFor(Promise.all(pendingQueries));
-  
-    // Assert
-    expect(seenAnswers).toEqual(expectedAnswers);
-  }))
-})`;
-
-const missingPartPBTSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), fc.array(fc.integer({min: 1, max: 10}), {minLength: 1}), async (s, batches) => {
-    // Arrange
-    const pendingQueries = [];
-    const seenAnswers = [];
-    const expectedAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-    const scheduledCall = s.scheduleFunction(call);
-    let concurrentQueriesDetected = false;
-    let queryPending = false;
-    const monitoredScheduledCall = (...args) => {
-      concurrentQueriesDetected ||= queryPending;
-      queryPending = true;
-      return scheduledCall(...args).finally(() => (queryPending = false));
-    };
-  
-    // Act
-    const queued = queue(monitoredScheduledCall);
-    let lastId = 0;
-    const { task } = s.scheduleSequence(batches.map((batch, index) => {
-      return {
-        label: \`Fire batch #\${index + 1} (\${batch} calls)\`,
-        builder: async () => {
-          for (let id = 0 ; id !== batch ; ++id, ++lastId) {
-            expectedAnswers.push(lastId);
-            pendingQueries.push(queued(lastId).then(v => (seenAnswers.push(v))));
-          }
-        },
-      }
-    }));
-    await s.waitFor(task);
-    await s.waitFor(Promise.all(pendingQueries));
-  
-    // Assert
-    expect(seenAnswers).toEqual(expectedAnswers);
-    expect(concurrentQueriesDetected).toBe(false);
-  }))
-})`;
-
-const extendedBackToWaitAllPBTSpecCode = `import {queue} from './queue.js';
-import fc from 'fast-check';
-
-test('should resolve in call order', async () => {
-  await fc.assert(fc.asyncProperty(fc.scheduler(), fc.array(fc.integer({min: 1, max: 10}), {minLength: 1}), async (s, batches) => {
-    // Arrange
-    const seenAnswers = [];
-    const expectedAnswers = [];
-    const call = jest.fn()
-      .mockImplementation(v => Promise.resolve(v));
-    const scheduledCall = s.scheduleFunction(call);
-    let concurrentQueriesDetected = false;
-    let queryPending = false;
-    const monitoredScheduledCall = (...args) => {
-      concurrentQueriesDetected ||= queryPending;
-      queryPending = true;
-      return scheduledCall(...args).finally(() => (queryPending = false));
-    };
-  
-    // Act
-    const queued = queue(monitoredScheduledCall);
-    let lastId = 0;
-    s.scheduleSequence(batches.map((batch, index) => {
-      return {
-        label: \`Fire batch #\${index + 1} (\${batch} calls)\`,
-        builder: async () => {
-          for (let id = 0 ; id !== batch ; ++id, ++lastId) {
-            expectedAnswers.push(lastId);
-            queued(lastId).then(v => (seenAnswers.push(v)));
-          }
-        },
-      }
-    }));
-    await s.waitAll();
-  
-    // Assert
-    expect(seenAnswers).toEqual(expectedAnswers);
-    expect(concurrentQueriesDetected).toBe(false);
-  }))
-})`;
-
-const extendedWithExceptionsPBTSpecCode = `import {queue} from './queue.js';
-import fc from "fast-check";
-
-test("should resolve in call order", async () => {
-  await fc.assert(
-    fc.asyncProperty(
-      fc.scheduler(),
-      fc.array(fc.integer({ min: 1, max: 10 }), { minLength: 1 }),
-      fc.func(fc.boolean()),
-      async (s, batches, isFailure) => {
-        // Arrange
-        const seenAnswers = [];
-        const expectedAnswers = [];
-        const call = jest
-          .fn()
-          .mockImplementation((v) =>
-            isFailure(v) ? Promise.reject(v) : Promise.resolve(v)
-          );
-        const scheduledCall = s.scheduleFunction(call);
-        let concurrentQueriesDetected = false;
-        let queryPending = false;
-        const monitoredScheduledCall = (...args) => {
-          concurrentQueriesDetected ||= queryPending;
-          queryPending = true;
-          return scheduledCall(...args).finally(() => (queryPending = false));
-        };
-
-        // Act
-        const queued = queue(monitoredScheduledCall);
-        let lastId = 0;
-        s.scheduleSequence(
-          batches.map((batch, index) => {
-            return {
-              label: \`Fire batch #\${index + 1} (\${batch} calls)\`,
-              builder: async () => {
-                for (let id = 0; id !== batch; ++id, ++lastId) {
-                  expectedAnswers.push(
-                    isFailure(lastId)
-                      ? \`failure:\${lastId}\`
-                      : \`success:\${lastId}\`
-                  );
-                  queued(lastId).then(
-                    (v) => seenAnswers.push(\`success:\${v}\`),
-                    (v) => seenAnswers.push(\`failure:\${v}\`)
-                  );
-                }
-              }
-            };
-          })
-        );
-        await s.waitAll();
-
-        // Assert
-        expect(seenAnswers).toEqual(expectedAnswers);
-        expect(concurrentQueriesDetected).toBe(false);
-      }
-    )
-  );
-})`;
+import {
+  queueCodeV0,
+  queueCodeV1,
+  queueCodeV2,
+  queueCodeV3,
+  queueCodeV4,
+  queueCodeV5,
+  queueUnitSpecCode,
+  queueBasicPBTSpecCode,
+  queueBasicPBTWaitAllSpecCode,
+  queueMoreThan2CallsPBTSpecCode,
+  queueFromBatchesPBTSpecCode,
+  queueBatchesAlternativePBTSpecCode,
+  missingPartPBTSpecCode,
+  extendedBackToWaitAllPBTSpecCode,
+  extendedWithExceptionsPBTSpecCode,
+} from './snippets.mjs';
 
 function SetupPlayground(props) {
   const { startSpecCode, anwserSpecCode, fileContent, fileName, fileExtension } = props;
