+++ website/jest.config.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  testMatch: ['<rootDir>/**/*.spec.mjs'],
+  testPathIgnorePatterns: ['/generated-tests/'],
+};
