+++ website/package.json
@@ -10,7 +10,8 @@
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
-    "write-heading-ids": "docusaurus write-heading-ids"
+    "write-heading-ids": "docusaurus write-heading-ids",
+    "test": "yarn node --experimental-vm-modules $(yarn bin jest)"
   },
   "dependencies": {
     "@codesandbox/sandpack-react": "^2.6.1",
@@ -25,7 +26,10 @@
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "2.4.0"
+    "@docusaurus/module-type-aliases": "2.4.0",
+    "@jest/globals": "^29.5.0",
+    "fast-check": "workspace:*",
+    "jest": "^29.5.0"
   },
   "browserslist": {
     "production": [
